requirements.txt
sentence-transformers==3.4.0
langchain 
flask 
pypdf
python-dotenv 
pinecone[grpc]
langchain-pinecone 
langchain_community 
langchain_openai 
langchain_experimental
-e .